fix(stars): resolve final payment and monitoring findings
