--- .github/workflows/pr310-review-fixes.yml
@@ -1,78 +1,0 @@
-name: Apply remaining PR 310 safety fixes
-
-on:
-  pull_request:
-    branches:
-      - main
-    types:
-      - opened
-      - reopened
-      - synchronize
-
-permissions:
-  contents: write
-
-concurrency:
-  group: pr310-final-fixes
-  cancel-in-progress: true
-
-jobs:
-  patch-test-commit:
-    if: github.event.pull_request.number == 310 && github.actor != 'github-actions[bot]'
-    runs-on: ubuntu-latest
-    timeout-minutes: 20
-    steps:
-      - uses: actions/checkout@v4
-        with:
-          ref: hotfix/stars-payment-safety
-          fetch-depth: 0
-
-      - uses: actions/setup-node@v4
-        with:
-          node-version: 20
-          cache: npm
-
-      - name: Apply current review fixes
-        run: python3 scripts/pr310-final-fixes.py
-
-      - name: Prepare writable data directory
-        run: |
-          sudo mkdir -p /data
-          sudo chmod 777 /data
-
-      - name: Install dependencies
-        run: npm ci --legacy-peer-deps
-
-      - name: Build TypeScript
-        run: npm run build
-
-      - name: Run full Jest suite
-        id: jest
-        continue-on-error: true
-        shell: bash
-        run: timeout 300s npm test -- --runInBand --forceExit > jest.log 2>&1
-
-      - name: Upload Jest diagnostics
-        if: steps.jest.outcome == 'failure'
-        uses: actions/upload-artifact@v4
-        with:
-          name: pr310-jest-log
-          path: jest.log
-          retention-days: 1
-
-      - name: Stop after Jest failure
-        if: steps.jest.outcome == 'failure'
-        run: |
-          tail -n 200 jest.log
-          exit 1
-
-      - name: Commit validated fixes
-        shell: bash
-        run: |
-          rm -f .github/workflows/pr310-review-fixes.yml
-          rm -f scripts/pr310-final-fixes.py
-          git config user.name "chatgpt-codex-connector[bot]"
-          git config user.email "199175422+chatgpt-codex-connector[bot]@users.noreply.github.com"
-          git add -A
-          git commit -m "fix(stars): resolve final payment and monitoring findings"
-          git push origin HEAD:hotfix/stars-payment-safety
