--- scripts/pr310-final-fixes.py
@@ -1,244 +1,0 @@
-from pathlib import Path
-import re
-
-
-def read(path: str) -> str:
-    return Path(path).read_text()
-
-
-def write(path: str, text: str) -> None:
-    Path(path).write_text(text)
-
-
-def one(path: str, old: str, new: str) -> None:
-    text = read(path)
-    count = text.count(old)
-    if count != 1:
-        raise RuntimeError(f"{path}: expected one match, found {count}: {old[:120]!r}")
-    write(path, text.replace(old, new, 1))
-
-
-def all_matches(path: str, old: str, new: str, expected: int) -> None:
-    text = read(path)
-    count = text.count(old)
-    if count != expected:
-        raise RuntimeError(f"{path}: expected {expected} matches, found {count}: {old[:120]!r}")
-    write(path, text.replace(old, new))
-
-
-def regex_one(path: str, pattern: str, replacement: str) -> None:
-    text = read(path)
-    updated, count = re.subn(pattern, replacement, text, count=1, flags=re.S)
-    if count != 1:
-        raise RuntimeError(f"{path}: expected one regex match, found {count}: {pattern[:120]!r}")
-    write(path, updated)
-
-
-# Existing paid Stars obligations must recover even when new checkout mode is BTC.
-one(
-    "src/services/stars-payment.ts",
-    "async function recoverPaidBundles(): Promise<void> {\n  if (recoveryRunning || !botInstance || !isStarsMode()) return;",
-    "export async function recoverPaidBundles(): Promise<void> {\n  if (recoveryRunning || !botInstance) return;",
-)
-
-# Monitoring may partially deliver an album. Record confirmed IDs and retry only the rest.
-one(
-    "src/controllers/send-active-stories.ts",
-    "    const partialIds = [...deliveredStoryIds];\n    if (partialIds.length > 0) {\n      throw new PartialStoryDeliveryError(error, partialIds);\n    }\n    throw error;",
-    "    const partialIds = [...deliveredStoryIds];\n    if ((task as any).monitorDelivery && partialIds.length > 0) {\n      console.warn(`[sendActiveStories] Partial monitor delivery for ${task.link}; remaining stories will retry.`);\n      return partialIds;\n    }\n    if (partialIds.length > 0) {\n      throw new PartialStoryDeliveryError(error, partialIds);\n    }\n    throw error;",
-)
-
-# Fold paid monitoring into the existing stoppable scheduler and bypass pagination.
-one(
-    "src/services/monitor-service.ts",
-    "      if (!premium && Number(monitor.telegram_id) !== BOT_ADMIN_ID) {\n        removeMonitor(monitor.telegram_id, monitor.target_id);\n        continue;\n      }",
-    "      const starsEntitlement = getStarsMonitoringEntitlement(monitor.telegram_id);\n      if (!premium && Number(monitor.telegram_id) !== BOT_ADMIN_ID && !starsEntitlement) {\n        removeMonitor(monitor.telegram_id, monitor.target_id);\n        continue;\n      }",
-)
-all_matches(
-    "src/services/monitor-service.ts",
-    "            initTime: Date.now(),\n",
-    "            initTime: Date.now(),\n            monitorDelivery: true,\n",
-    2,
-)
-
-# Remove the independent Stars monitor timer; /stopmonitor now stops all monitoring.
-one(
-    "src/services/stars-mode-safety.ts",
-    "const MONTH_SECONDS = 30 * 24 * 60 * 60;\nconst MONITOR_INTERVAL_MS = 60 * 60 * 1000;\n\nlet monitorTimer: NodeJS.Timeout | null = null;\nlet monitorCycleRunning = false;\nlet launchPatched = false;",
-    "const MONTH_SECONDS = 30 * 24 * 60 * 60;\n\nlet launchPatched = false;",
-)
-regex_one(
-    "src/services/stars-mode-safety.ts",
-    r"\nasync function runStarsMonitorCycle\(\): Promise<void> \{.*?\nfunction retainPendingTelegramUpdates",
-    "\nfunction retainPendingTelegramUpdates",
-)
-one(
-    "src/services/stars-mode-safety.ts",
-    "  retainPendingTelegramUpdates(bot);\n  startStarsMonitorLoop();",
-    "  retainPendingTelegramUpdates(bot);",
-)
-
-# Restoring a lower monitoring limit must remove excess active rows immediately.
-one(
-    "src/services/stars-mode-safety.ts",
-    "      UPDATE star_monitor_grants\n      SET refunded_at = NEW.refunded_at\n      WHERE bundle_id = NEW.bundle_id AND refunded_at IS NULL;\n\n      DELETE FROM star_monitor_entitlements",
-    """      UPDATE star_monitor_grants
-      SET refunded_at = NEW.refunded_at
-      WHERE bundle_id = NEW.bundle_id AND refunded_at IS NULL;
-
-      INSERT OR REPLACE INTO star_monitor_delete_authorizations (
-        telegram_id, target_id, authorized_at
-      )
-      SELECT m.telegram_id, m.target_id, CAST(strftime('%s','now') AS INTEGER)
-      FROM monitors m
-      JOIN star_monitor_entitlements e ON e.user_id = m.telegram_id
-      WHERE e.user_id = (
-        SELECT user_id FROM star_monitor_grants WHERE bundle_id = NEW.bundle_id
-      )
-      AND (
-        SELECT COUNT(*) FROM monitors earlier
-        WHERE earlier.telegram_id = m.telegram_id
-          AND (
-            COALESCE(earlier.created_at, 0) < COALESCE(m.created_at, 0)
-            OR (
-              COALESCE(earlier.created_at, 0) = COALESCE(m.created_at, 0)
-              AND earlier.id <= m.id
-            )
-          )
-      ) > e.max_targets;
-
-      DELETE FROM monitors
-      WHERE telegram_id = (
-        SELECT user_id FROM star_monitor_grants WHERE bundle_id = NEW.bundle_id
-      )
-      AND (
-        SELECT COUNT(*) FROM monitors earlier
-        WHERE earlier.telegram_id = monitors.telegram_id
-          AND (
-            COALESCE(earlier.created_at, 0) < COALESCE(monitors.created_at, 0)
-            OR (
-              COALESCE(earlier.created_at, 0) = COALESCE(monitors.created_at, 0)
-              AND earlier.id <= monitors.id
-            )
-          )
-      ) > COALESCE(
-        (SELECT max_targets FROM star_monitor_entitlements e WHERE e.user_id = monitors.telegram_id),
-        0
-      );
-
-      DELETE FROM star_monitor_delete_authorizations
-      WHERE telegram_id = (
-        SELECT user_id FROM star_monitor_grants WHERE bundle_id = NEW.bundle_id
-      );
-
-      DELETE FROM star_monitor_entitlements""",
-)
-one(
-    "src/services/stars-mode-safety.ts",
-    "      AND NOT EXISTS (\n        SELECT 1\n        FROM star_monitor_grants g\n        WHERE g.user_id = star_monitor_entitlements.user_id\n          AND g.refunded_at IS NULL\n      );\n    END;",
-    """      AND NOT EXISTS (
-        SELECT 1
-        FROM star_monitor_grants g
-        WHERE g.user_id = star_monitor_entitlements.user_id
-          AND g.refunded_at IS NULL
-      );
-
-      DELETE FROM monitors
-      WHERE telegram_id = (
-        SELECT user_id FROM star_monitor_grants WHERE bundle_id = NEW.bundle_id
-      )
-      AND NOT EXISTS (
-        SELECT 1 FROM star_monitor_entitlements e
-        WHERE e.user_id = monitors.telegram_id
-          AND e.expires_at > CAST(strftime('%s','now') AS INTEGER)
-      );
-    END;""",
-)
-
-# Clear legacy chat-wide group menus, then continue with member-scoped menus.
-one(
-    "src/services/stars-command-surface.ts",
-    "const COMMAND_SCOPE_MIGRATION_KEY = 'stars_command_scope_v3';",
-    "const COMMAND_SCOPE_MIGRATION_KEY = 'stars_command_scope_v4';",
-)
-one(
-    "src/services/stars-command-surface.ts",
-    "  const users = db.prepare(\n    'SELECT telegram_id, language FROM users ORDER BY created_at ASC',\n  ).all() as { telegram_id: string; language?: string }[];",
-    """  const legacyGroups = db.prepare(`
-    SELECT DISTINCT group_id
-    FROM (
-      SELECT CAST(telegram_id AS TEXT) AS group_id
-      FROM download_queue
-      WHERE CAST(telegram_id AS INTEGER) < 0
-      UNION
-      SELECT CAST(json_extract(task_details, '$.chatId') AS TEXT) AS group_id
-      FROM download_queue
-      WHERE json_valid(task_details)
-        AND CAST(json_extract(task_details, '$.chatId') AS INTEGER) < 0
-      UNION
-      SELECT CAST(chat_id AS TEXT) AS group_id
-      FROM star_result_bundles
-      WHERE CAST(chat_id AS INTEGER) < 0
-    )
-    WHERE group_id IS NOT NULL
-    ORDER BY group_id
-  `).all() as { group_id: string }[];
-
-  for (const row of legacyGroups) {
-    const groupId = Number(row.group_id);
-    if (!Number.isFinite(groupId)) continue;
-    try {
-      await (bot.telegram as any).callApi('deleteMyCommands', {
-        scope: { type: 'chat', chat_id: groupId },
-      });
-    } catch (error) {
-      console.warn(`[Stars] Could not clear legacy group command scope ${row.group_id}:`, error);
-    }
-    await new Promise((resolve) => setTimeout(resolve, 100));
-  }
-
-  const users = db.prepare(
-    'SELECT telegram_id, language FROM users ORDER BY created_at ASC',
-  ).all() as { telegram_id: string; language?: string }[];""",
-)
-
-Path("__tests__/stars-final-safety.test.ts").write_text("""import fs from 'fs';
-
-const source = (path: string) => fs.readFileSync(path, 'utf8');
-
-describe('final Stars safety invariants', () => {
-  test('paid recovery is independent of checkout mode', () => {
-    const text = source('src/services/stars-payment.ts');
-    expect(text).toContain('export async function recoverPaidBundles');
-    expect(text).toContain('if (recoveryRunning || !botInstance) return;');
-    expect(text).not.toContain('recoveryRunning || !botInstance || !isStarsMode()');
-  });
-
-  test('paid and monitored active delivery bypass pagination and preserve partial IDs', () => {
-    const text = source('src/controllers/send-active-stories.ts');
-    expect(text).toContain('task.starsBundleId || (task as any).monitorDelivery');
-    expect(text).toContain('return partialIds');
-    expect(text).toContain('new PartialStoryDeliveryError(error, partialIds)');
-  });
-
-  test('paid monitoring uses the existing stoppable scheduler', () => {
-    const monitor = source('src/services/monitor-service.ts');
-    const safety = source('src/services/stars-mode-safety.ts');
-    expect(monitor).toContain('getStarsMonitoringEntitlement(monitor.telegram_id)');
-    expect(monitor.match(/monitorDelivery: true/g)?.length).toBe(2);
-    expect(safety).not.toContain('startStarsMonitorLoop');
-    expect(safety).not.toContain('runStarsMonitorCycle');
-  });
-
-  test('refunds reconcile monitor rows and command migration clears group scopes', () => {
-    const safety = source('src/services/stars-mode-safety.ts');
-    const commands = source('src/services/stars-command-surface.ts');
-    expect(safety).toContain('DELETE FROM monitors');
-    expect(commands).toContain('stars_command_scope_v4');
-    expect(commands).toContain("callApi('deleteMyCommands'");
-    expect(commands).toContain("type: 'chat_member'");
-  });
-});
-""")
-
-print("Applied remaining PR 310 safety fixes.")
